Add support for 64-bit AKF mailboxes
AKF (ARM KingFisher, i.e. Cortex-A7) are a class of mailboxes
used for Cortex-A7 coprocessors in Apple SoCs.

The filename is actually a bit of a misnomer: The added files deal with
the 64-bit variant of AKF mailbox, which are present on Apple A7-A10 SoCs.

Apple A11, T2 also have AKF mailboxes, but they are of a 96-bit interface
that is compatible with the existing ASC mailbox interface, so it is
handled there instead.

Since the message size of this mailbox is 64-bit instead of 96-bit,
support for this type of mailbox in its own file as all clients will need
to be modified to support the smaller message size.

Signed-off-by: Nick Chan <towinchenmi@gmail.com>