Add support for AKF mailboxes
AKF (Apple KingFisher) mailboxes are used for Cortex-A7 coprocessors
in A7 - A10, T2 SoCs. There are two revisions of this hardware. Since the
message size of this mailbox is 64-bit instead of 96-bit, support this
type of mailbox in its own file as all clients will need to be modified to
support the smaller message size.

Signed-off-by: Nick Chan <towinchenmi@gmail.com>